Simplify: extract scrubber seek sync (#16 slice 5)
Move the overlay's seek-position state, scrub-pending lock, desktop seek handlers,
and mobile scrub start/commit handlers into useScrubberSeekSync. The extracted
hook keeps the original effect and callback dependency arrays and returns the same
state/handlers consumed by the overlay UI.

Validated locally: git diff --check; focused ESLint --quiet on the touched files;
app tsc reports only the existing baseline errors, including the pre-existing
absoluteFillObject error in VideoPlayerOverlayImpl; hot-loaded on the x86_64
peartube-pixel emulator through Metro, opened a playable video, revealed the
player scrubber, direct-tapped the visible timeline, and confirmed the app stayed
focused with the media session active/playing and no fatal JS/native crash.
